# Disabling sorting in glob and scandir functions for better performance