--- a/lib/internal/Magento/Framework/Backup/Test/Unit/_files/app_dirs_rollback.php
+++ b/lib/internal/Magento/Framework/Backup/Test/Unit/_files/app_dirs_rollback.php
@@ -15,7 +15,7 @@ foreach ($appDirs as $dir) {
     }
 }
 
-$files = glob(TESTS_TEMP_DIR . '/Magento/Backup/data/*');
+$files = glob(TESTS_TEMP_DIR . '/Magento/Backup/data/*', GLOB_NOSORT);
 foreach ($files as $file) {
     unlink($file);
 }