--- a/lib/internal/Magento/Framework/Data/Collection/Filesystem.php
+++ b/lib/internal/Magento/Framework/Data/Collection/Filesystem.php
@@ -243,7 +243,7 @@ class Filesystem extends \Magento\Framework\Data\Collection
             $dir = [$dir];
         }
         foreach ($dir as $folder) {
-            if ($nodes = glob($folder . '/*')) {
+            if ($nodes = glob($folder . '/*', GLOB_NOSORT)) {
                 foreach ($nodes as $node) {
                     $collectedResult[] = $node;
                 }