Disabling sorting in glob and scandir functions for better performance

### Description
Disabling sorting in glob and scandir functions for better performance

https://www.exakat.io/php-likes-sorting/