@magento-engcom-team 2018-06-12
Hi @orlangur, thank you for the review.
[ENGCOM-1972](https://jira.corp.magento.com/browse/ENGCOM-1972) has been created to process this Pull Request