@gamort 2018-10-03
This was a really really REALLY bad idea.
Specifically for 
lib/internal/Magento/Framework/App/Utility/Files.php
 lib/internal/Magento/Framework/Data/Collection/Filesystem.php
and
 lib/internal/Magento/Framework/Filesystem/Io/File.php

 The rest of the changes in this PR were fine - they are made to functions were there was no expectation of sorting.   I doubt it saved more than 100ms for unit testing and deployment, but that is ok.   The problem is forcing a non-standard default behaviour in general purpose classes.  And it is forced - there is no way to actually turn sorting back on AND use the file access classes.

The benchmark from https://www.exakat.io/php-likes-sorting/ was irrelevant.   That benchmark was for sorting 26,000 files that matched the pattern to begin with.  Your not sorting thousands of files, at most hundreds, more likely dozens.

Example:  A magento installation with 100,000 files, 260 of them are xml files
A glob search will first locate 260 xml files.  It will THEN sort 260 xml files.   So you are not going from 14s to 12s, more like 140ms to 120ms.  Saving 20ms was a useless waste of time.

The default behaviour of glob searches is to sort.   Everyone expects them to sort.   Short of using the glob function directly[and therefore losing all the benefits of a standardized file system set of classes] there is no way to sort other then to sort the array of files which are returned.   Which will add 200ms or more since array sorting in PHP is less effiicient! [See the same article above].

Now, WHY does everyone want sorting to take place by default?

Simple, whenever we operate on a list of files, we don' have to provide any advanced "work on file X before working on file Y" programming logic.   We just name our files appropriately.   Look in your php/conf.d folder and you will immediately see it:
10-somext.ini
20-someext.ini
30-somext.ini
zz-magento.ini
The magento specific PHP configuration settings are always applied last. 

Why do we do this in a very low level system function instead of higher level code?  Because it is stupid to optimize hundreds of thousands of applications when you can optimize a single system function.   And hey, YOU don't have to optimize it, thousands of programmers have spent decades optimizing it for you.